requirements.txt
opencv-python
lxml
tqdm
tensorflow
absl-py
easydict
matplotlib
pillow
tensorflow_addons
filterpy
numba
scikit-learn==0.22.2